Makefile: use "ln -f" instead of "rm && ln"
Change the invocations and behavior of "ln-or-cp.sh" to not assume
that we're going to "rm" the file in question beforehand.

This reduces the complexity of these rules, and as a bonus means we'll
likely have a smaller race in a case where we want to "make install"
onto a running system.

That'll still need to be implemented in terms of an unlink(2) followed
by a link(2) or symlink(2), so the race is still there, but it's
probably smaller than what our Makefile will produce. To truly
eliminate that race we'd need to create the symlink elsewhere and
rename it in place.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>