Makefile: add a INSTALL_FALLBACK_LN_CP mode
Change the default behavior on "make install" where we fallback
through a chain of "ln || ln -s || cp" to instead error out when we
can't symlink or hardlink, and not then fallback on a "cp" (or from a
symlink to hardlink etc.).

This does not change the fallback behavior that the user can ask for,
but it does change silently degradingthrough a chain of "ln || ln -s
|| cp" into an error by default. So e.g. someone who wants hardlinks
normally, but can't have them between /usr/bin and
/usr/libexec (different FS's) will now need:

     make NO_CROSS_DIRECTORY_HARDLINKS=Y install

Someone who's e.g. building packages of git is much better of knowing
that their settings couldn't work, than e.g. having things degrade
down to "cp", only to find that their git package ballooned to some
ridiculous size because it didn't use hardlinks.

The previous fallback behavior was introduced in
3e073dc5611 (Makefile: always provide a fallback when hardlinks fail,
2008-08-25), since then we've gained the ability to specify e.g. that
we'd like symlinks via the INSTALL_SYMLINKS setting.

This also changes the behavior for the build (not install!) to always
use the new strict mode. This will theoretically break things for
someone who can't make symlinks or hardlinks in their git checkout
when building.

That part of this change would break building on Windows and other
platforms that don't support symlinks if INSTALL_SYMLINKS were to
become the default, but it's not on by default, so we should be fine
here. If and when INSTALL_SYMLINKS ever becomes the default the right
way to deal with this would be to tweak config.mak.uname
appropriately, not to silently fall back on a copy.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>