send-email: lazily load config for a big speedup
Reduce the time it takes git-send-email to get to even the most
trivial of tasks (such as serving up its "-h" output) by first listing
config keys that exist, and only then only call e.g. "git config
--bool" on them if they do.

Over a lot of runs this speeds the time to "-h" up for me from ~250ms
to ~150ms, and the runtime of t9001-send-email.sh goes from ~25s to
~20s.

This introduces a race condition where we'll do the "wrong" thing if a
config key were to be inserted between us discovering the list and
calling read_config(). In this case that race condition doesn't
matter.

The config_regexp() function was added in dd84e528a34 (git-send-email:
die if sendmail.* config is set, 2020-07-23) for use by
git-send-email, so we can change its odd return value in the case
where no values are found by git-config(1). The difference in the *.pm
code would matter if it was invoked in scalar context, but now it no
longer is.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>