send-email: move trivial config handling to Perl
Optimize the startup time of git-send-email by using an amended
config_regexp() function to retrieve the list of config keys and
values we're interested in.

For boolean keys we can handle the [true|false] case ourselves, and
the "--get" case didn't need any parsing. Let's leave "--path" and
other "--bool" cases to git-config(1).

The config_regexp() function was added in dd84e528a34 (git-send-email:
die if sendmail.* config is set, 2020-07-23) for use by
git-send-email, so we can just change it.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>