test-lib tests: grep out verbose "git init" output
Grep out the output emitted by "git init" when not under the "--quiet"
option, as is done when test-lib.sh itself invokes it, and when "git
init" or "test_create_repo()" do.

The output contains the absolute path to the initialized repository,
so it can't be test_cmp'd without munging it. Let's instead ignore it
for now, change later down the line will make the output stable
instead.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>