SubmittingPatches: remove pine-specific hints from MUA hints
Replace the whole section about configuring pine with
no-strip-whitespace-before-send with a note that users should use
another MUA instead.

There hasn't been a release of Pine since late 2005, in Alpine the
no-strip-whitespace-before-send switch is already the default.

This documentation was initially added in 9740d289ee7 (Update
SubmittingPatches to add MUA specific notes., 2005-08-26). At this
point it's safe to say that it's not useful to anyone anymore, let's
remove it.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>