repository.h: don't use a mix of int and bitfields
Change the bitfield added in 58300f47432 (sparse-index: add
index.sparse config option, 2021-03-30) and 3964fc2aae7 (sparse-index:
add guard to ensure full index, 2021-03-30) to just use an "int"
boolean instead.

It might be smart to optimize the space here in the future, but for a
subsequent refactoring let's start out by making this consistent.

Additionally, since this was "unsigned" and we'd memset the structure
to -1 we'd initialize both of these to "1". So the default setting of:

    r->settings.command_requires_full_index = 1

In 3964fc2aae7 was redundant until now. The "Initialize this as off"
comment in 58300f47432 was also inaccurate, we'd initialized it as
"1", and then reset it to "0". We'll now change it from "-1" to "0" in
that case instead.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>