read-cache: fix incorrect count and progress bar stalling
Fix a potential incorrect display of the number of items (off by one)
and stalling of the progress bar in refresh_index().

The off-by-one error is minor, we should say we're processing the 1st
item, not the 0th. This along with the next change also allows us to
remove the last display_progress() call outside the loop, as we'll
always have reached 100% now.

The next change is to move the display_progress() call to the very
start of the loop. In the loop we first check whether e.g. we ignore
submodules and the entry we're processing is a submodule, whether we
ignore certain paths etc.. Thus we could have a pathological case
where we have a huge index consisting of such ignored entries, and
we'd stall on the progress bar.

Let's instead call display_progress() as the first thing we do in the
loop.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>