pack-objects: move builtin-only code to its own header
Move the code that's only used in builtin/pack-objects.c to a new
builtin/pack-objects.h header and out of pack-objects.h.

This fixes an issue where Solaris's SunCC hasn't been able to compile
git since 483fa7f42d9 (t/helper/test-bitmap.c: initial commit,
2021-03-31). The real origin of that issue is that in
898eba5e630 (pack-objects: refer to delta objects by index instead of
pointer, 2018-04-14) utility functions only needed by
builtin/pack-objects.c were added to pack-objects.h. Since then the
header has been used in a few other places, but 483fa7f42d9 was the
first time it was used by test helper.

Since Solaris is stricter about linking and the oe_get_size_slow()
function lives in builtin/pack-objects.c the build started failing
with:

    Undefined                       first referenced
     symbol                             in file
    oe_get_size_slow                    t/helper/test-bitmap.o
    ld: fatal: symbol referencing errors. No output written to t/helper/test-tool

On other platforms this is presumably OK because the compiler and/or
linker detects that the "static inline" functions that reference
oe_get_size_slow() aren't used.

Introducing the first *.h file in builtin/* is a somewhat odd solution
to this problem, but I think more palatable than moving this to
builtin/pack-objects.c directly. Since oe_get_size_slow() is present
later in that file we'd either need a forward definition of it (which
to be fair, we're doing in the header here, but were also before this
code move). To avoid the forward definition we'd need to move
oe_get_size_slow() to the top of the file, which would make for a
bigger change than this.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>