docs: start moving technical "format" docs to man pages
Move the Documentation/technical/bundle-format.txt documentation added
in 7378ec90e1c (doc: describe Git bundle format, 2020-02-07) to a new
gitformat-bundle(5) page, and add a gitformats(5) index page which
will be an index of such documentation. We'll move more of it over in
subsequent commits.

The reason to do this is that having this documentation in
Documentation/technical/* makes it less accessible. It is not
installed by default, and we don't build man page versions of these,
therefore when we link to it from man pages we end up linking to the
generated HTML version of these docs.

I don't see a reason for why most or all of of
Documentation/technical/* documentation of this nature shouldn't be
promoted to being a part of our main documentation, e.g. in this case
the git-bundle(1) manpage can now link to this new documentation.

Unfortunately the syntax of the current Documentation/technical/*.txt
is not the same (when it comes to section headings etc.) as our
Documentation/*.txt documentation, so change the relevant bits of
syntax as we're moving this over.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>