bundle tests: use test_cmp instead of grep
Change the bundle tests to fully compare the expected "git ls-remote"
or "git bundle list-heads" output, instead of merely grepping it.

This avoids subtle regressions in the tests. In
f62e0a39b6 (t5704 (bundle): add tests for bundle --stdin, 2010-04-19)
the "bundle --stdin <rev-list options>" test was added to make sure we
didn't include the tag.

But since the --stdin mode didn't work until 5bb0fd2cab (bundle:
arguments can be read from stdin, 2021-01-11) our grepping of
"master" (later "main") missed the important part of the test, namely
that we should not include the "refs/tags/tag" tag in that case. Since
the test only grepped for "main" in the output we'd miss a regression
in that code.

So let's use test_cmp instead, and also in the other nearby tests
where it's easy.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>