progress.c tests: add a "signal" verb
Add a "signal" synonym for "update". It is not typical of the
progress.c API to encounter a scenario where we do an update before
the first display_progress(), a second will need to have passed
between start_progress() and the first call to display_progress() for
that to happen.

Let's make the tests more self-documenting by indicating this state
explicitly by calling such instances "signal". Under the hood this is
simply a synonym for "update", but we can imagine that later "update"
calls would elide many display_progress() calls.

Nothing makes use of this right now, but in a future commit I expect
to start treating this these "signal" cases differently, i.e. to use
the signal as a trigger to show a "stalled?" state to the user if
we've gone the full second after start_progress() without calling a
display_progress().

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>