progress.c: add a "global_progress" variable
It's a current limitation of the progress.c bar code that any update
must pro-actively go through the likes of display_progress().

If we e.g. hang forever before the first display_progress(), or in the
middle of a loop that would call display_progress() the user will only
see either no output, or output frozen at the last display_progress()
that would have done an update (e.g. in cases where progress_update
was "1" due to an earlier signal).

This change does not fix that, but sets up the structure for solving
that and other problems by juggling a "global_progress" struct. We do
not support having more than one of these at a time anyway in terms of
the output we emit, so it's good to have this assertion.

Subsequent commits will make more use of the "global_progress" than
only using it for these assertions.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>